Refactor DTO and wrapper classes to packages
