+++ moviecatalog/src/main/java/com/drbsimon/moviecatalog/model/dto/SeatDTO.java
@@ -1,4 +1,4 @@
-package com.drbsimon.moviecatalog.model;
+package com.drbsimon.moviecatalog.model.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,9 +9,9 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Seat {
+public class SeatDTO {
     private Long id;
     private Integer rowNumber;
     private Integer seatNumber;
-    private Room room;
+    private RoomDTO room;
 }
