+++ moviecatalog/src/main/java/com/drbsimon/moviecatalog/model/wrapper/RoomListWrapper.java
@@ -1,5 +1,6 @@
-package com.drbsimon.moviecatalog.model;
+package com.drbsimon.moviecatalog.model.wrapper;
 
+import com.drbsimon.moviecatalog.model.dto.RoomDTO;
 import lombok.Data;
 import org.springframework.stereotype.Component;
 
@@ -8,5 +9,5 @@ import java.util.List;
 @Component
 @Data
 public class RoomListWrapper {
-    List<Room> rooms;
+    List<RoomDTO> rooms;
 }
